CameraController: Explicitly use ImageReader.setImageFormat() with YUV stream
When ImageReader.Builder is used which CameraController uses, the
default dataspace is set to UNKNOWN instead of JFIF. JFIF is the
default dataspace when ImageReader is created without a builder. The UNKNOWN
dataspace with a YUV stream can confuse some camera HALs. So instead, use
ImageReader.setImageFormat(), which derives the dataspace internally.

Bug: 393836416

Test: manual

Flag: EXEMPT; bugfix

Change-Id: I41c8b7f6435d607021b279516118c9b2c1c541a4
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>