Reduce HQ button size back to 32dp
The size of the button used to togglig High Quality was
increased in Ib5e40dc0cf to adhere to accessibility guidelines that
state that touch targets must have aa minimum size of 40dp. However
the larger size looks aestetically worse.

As a compromise, this CL reduces the HQ button back to the original
size, but increases the touch target of the HQ button to be at least
48dp..

Bug: 375117093
Test: Validated with Accessibility scanner that the touch target size
      is now 48dp, and does not raise a warning.
Flag: EXEMPT bugfix
Change-Id: I560f6d05347940d4201f1ecfe517a250badc4d4d