Accessibility: Don't disable HQ toggle on press
The current logic disables the High Quality button on touch
until WebcamController toggles HQ. This is done to debouce fast
presses as the WebcamController may take a little while to toggle
HQ mode on or off. The button is re-enabled when WebcamController
reports that HQ mode had been toggled.

Unfortunately, with the current logic, Screen Readers will announce
"Disabled" or "Enabled" every time the button is disabled/enabled,
which can be confusing to users.

This CL adds a sentinel to debounce fast toggle requests and does away
with the button enable/disable logic. The button simply no-ops if
WebcamController is currently in the middle of  toggling HQ mode.

Bug: 375299370
Test: Verified that Talkback does not announce "Enabled" or "Disabled"
Flag: EXEMPT bugfix
Change-Id: Ibf235f943e0d6661ddf38bc27410135a4efd64d6