Accessibility: Use sp for zoom controller's text size units.
The current dimension used in ZoomController is specified in
dp, which scales with the user's display size, not the text size.
This means that the zoom values did not scale when the users increased
or decreased their system wide text scaling.

This CL changes the units to sp, which scales with user's chosen text
scaling, instead of display scaling.

Bug: 386737747
Test: Verified that the zoom ratio text changes with text scaling.
Flag: EXEMPT bugfix.
Change-Id: I566aacdbda9a9ce2fb3703ef0a24c847f64c5930